refactor: routes comments 모듈화
Related to: #143
